requirements.txt
dask[dataframe]
facebook-sdk
pandas
python-dotenv
